# config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Documentação API PetMap",
    "description": "Esta documentação serve como guia para desenvolvedores entenderem como utilizar e integrar a API PetMap de forma eficiente, detalhando endpoints, parâmetros e métodos de autenticação.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira o token no formato: Bearer {token}"
    }
  },
  "paths": {
...
    "/api/pet/location/all": {
      "get": {
        "tags": [
          "Pet"
        ],
        "summary": "Lista todas as localizações dos pets",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista de pets no mapa!",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "errorMessage": {
                  "type": "string",
                  "example": ""
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "lat": {
                        "type": "number",
                        "example": -42.23099433
                      },
                      "lng": {
                        "type": "number",
                        "example": -8.234699818
                      },
                      "petId": {
                        "type": "string",
                        "example": "Elsndsdasosd8w"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/pet/find/update": {
      "put": {
        "tags": [
          "Pet"
        ],
        "summary": "Alterar informações/localização do Pet.",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "petId": {
                  "example": "any"
                },
                "localicacao": {
                  "example": "any"
                },
                "descricao": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/pet/find/remove": {
      "delete": {
        "tags": [
          "Pet"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "petId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/pet/find/register": {
      "post": {
        "tags": [
          "Pet"
        ],
        "summary": "Endpoint de registro de um novo pet e registrar no mapa.",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "apelido": {
                  "example": "any"
                },
                "localizacao": {
                  "example": "any"
                },
                "descricao": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "coleira": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/pet/locateByLatLng": {
      "post": {
        "tags": [
          "Pet"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "lat": {
                  "example": "any"
                },
                "lng": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/pet/find/get/{id}": {
      "get": {
        "tags": [
          "Pet"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/pet/myPets": {
      "get": {
        "tags": [
          "Pet"
        ],
        "summary": "Endpoint para retornar os pets do usuário logado.",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/pet/find/search": {
      "get": {
        "tags": [
          "Pet"
        ],
        "summary": "Pesquisa por pets usando um único parâmetro.",
        "description": "",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}